review: Move `DateUtils` to `Core2`
