+++ Core2/src/main/java/com/infomaniak/core2/DateUtils.kt
@@ -15,7 +15,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package com.infomaniak.swisstransfer.ui.utils
+package com.infomaniak.core2
 
 import java.text.SimpleDateFormat
 import java.util.Date
